ppt-to-pdf-microservice/requirements.txt
fastapi
uvicorn
python-pptx
PyMuPDF
reportlab
